validated and fixed reason for missing schools
